feat: make sidebar collapsable
