+++ src/components/sidebar/Sidebar.module.css
@@ -1,0 +1,40 @@
+.sidebar {
+  display: flex;
+  flex-direction: column;
+  border-right: 2px solid var(--color-bg-secondary);
+  background-color: var(--color-bg);
+  box-sizing: border-box;
+  padding: 12px;
+  width: 300px;
+  height: 100%;
+}
+
+.logo {
+  text-align: center;
+  text-decoration: none;
+  color: var(--color-text);
+  font-weight: 600;
+}
+.logo:hover {
+  text-decoration: underline;
+}
+
+.bottom {
+  display: flex;
+  flex-direction: column;
+  margin-top: auto;
+}
+
+.collapsed {
+  width: fit-content;
+}
+
+@media screen and (max-width: 800px) {
+  .sidebar {
+    position: fixed;
+  }
+
+  .collapsed {
+    position: relative;
+  }
+}
