+++ src/components/sidebar/Sidebar.tsx
@@ -1,0 +1,67 @@
+import { useState } from 'react';
+import { Link, useNavigate } from 'react-router';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { useCurrentUser } from '@/hooks/useCurrentUser';
+import { requestLogout } from '@/api/user';
+import { ChevronsLeft, ChevronsRight, Home, LogOut, Settings } from 'lucide-react';
+import { SidebarButton } from '../ui/sidebar-button';
+import classes from './Sidebar.module.css';
+import clsx from 'clsx';
+
+export function Sidebar() {
+  const { currentUser } = useCurrentUser();
+
+  const navigate = useNavigate();
+  const queryClient = useQueryClient();
+
+  const { mutate: logout, isPending } = useMutation({
+    mutationFn: requestLogout,
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: ['currentUser'] });
+      navigate('/login');
+    },
+  });
+
+  const [collapsed, setCollapsed] = useState(window.innerWidth <= 800);
+
+  if (!currentUser) return <span>Loading...</span>;
+
+  return (
+    <nav className={clsx(classes.sidebar, collapsed && classes.collapsed)}>
+      <Link to="/">
+        {collapsed ? (
+          <SidebarButton icon={<Home size={20} />} />
+        ) : (
+          <span className={classes.logo}>{import.meta.env.VITE_APP_NAME}</span>
+        )}
+      </Link>
+
+      <div className={classes.bottom}>
+        {(currentUser.role === 'admin' || currentUser.role === 'teacher') && (
+          <Link to="/panel">
+            <SidebarButton icon={<Settings size={20} />}>
+              {collapsed ? '' : 'Панель управления'}
+            </SidebarButton>
+          </Link>
+        )}
+
+        <SidebarButton
+          variant="danger"
+          onClick={() => logout()}
+          showLoader={isPending}
+          className={classes.exit}
+          icon={<LogOut size={20} />}
+        >
+          {collapsed ? '' : 'Выйти'}
+        </SidebarButton>
+
+        <SidebarButton
+          icon={collapsed ? <ChevronsRight size={20} /> : <ChevronsLeft size={20} />}
+          onClick={() => setCollapsed(!collapsed)}
+        >
+          {collapsed ? '' : 'Свернуть'}
+        </SidebarButton>
+      </div>
+    </nav>
+  );
+}
