+++ src/components/ui/button/Button.module.css
@@ -5,23 +5,40 @@
   gap: 8px;
   font-size: 16px;
   font-weight: 500;
-  padding: 12px 16px;
-  border-radius: 4px;
   border: none;
   cursor: pointer;
   transition: background-color 200ms;
-  background-color: var(--color-primary);
-  color: var(--color-bg);
-}
-.button:hover {
-  background-color: color-mix(in srgb, var(--color-primary), #000 20%);
-}
-.button:active {
-  background-color: color-mix(in srgb, var(--color-primary), #000 40%);
 }
 .button:disabled {
   background-color: color-mix(in srgb, var(--color-primary), #000 40%);
   cursor: not-allowed;
+}
+
+.primary {
+  padding: 12px 16px;
+  border-radius: 4px;
+  background-color: var(--color-primary);
+}
+.primary:hover {
+  background-color: color-mix(in srgb, var(--color-primary), #000 20%);
+}
+.primary:active {
+  background-color: color-mix(in srgb, var(--color-primary), #000 40%);
+}
+.primary:disabled {
+  background-color: color-mix(in srgb, var(--color-primary), #000 40%);
+}
+
+.icon {
+  padding: 8px;
+  border-radius: 50px;
+  background-color: transparent;
+}
+.icon:hover {
+  background-color: var(--color-bg-secondary);
+}
+.icon:active {
+  background-color: color-mix(in srgb, var(--color-bg-secondary), #000 10%);
 }
 
 .loader {
