+++ src/layouts/authorized-layout/AuthorizedLayout.module.css
@@ -1,41 +1,10 @@
 .container {
   display: flex;
   height: 100%;
-  gap: 24px;
-}
-
-.sidebar {
-  display: flex;
-  flex-direction: column;
-  border-right: 2px solid var(--color-bg-secondary);
-  box-sizing: border-box;
-  padding: 12px;
-  width: 300px;
-  height: 100%;
 }
 
 .main {
   flex: 1;
   overflow-y: auto;
-}
-
-.logos {
-  display: flex;
-  flex-direction: column;
-  font-weight: 600;
-  font-size: 14px;
-  color: var(--color-text);
-  text-decoration: none;
-  padding: 8px;
-  border-radius: 4px;
-  transition: background-color 200ms ease;
-}
-.logos:hover {
-  background-color: var(--color-bg-secondary);
-}
-
-.bottom {
-  display: flex;
-  flex-direction: column;
-  margin-top: auto;
+  padding: 12px;
 }
